backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite==0.20.0
alembic==1.13.1
pydantic[email]
pydantic-settings
python-jose[cryptography]==3.3.0
bcrypt
python-multipart==0.0.9
langchain
langchain-groq
langchain-core
python-dotenv==1.0.1
httpx